// src/models/select-picker-settings.ts
export interface SelectPickerSettings {
  emptyText: string;
  selectedCountText: string;
  wrapperClass: string;
  buttonClass: string;
  buttonPlaceholderClass: string;
  dropdownClass: string;
  searchWrapperClass: string;
  searchInputClass: string;
  searchInputPlaceholder: string;
  listBoxClass: string;
  optionClass: string;
  tickClass: string;
  tickContent: string;
  tickHiddenClass: string;
  selectedOptionClass: string;
}